# backend-api-spec.yaml
openapi: 3.0.3
info:
  title: CityPulse Energy Management API
  description: Comprehensive API for CityPulse Energy Management System
  version: 1.0.0
  contact:
    name: Marc (Backend Developer)
    email: marc@citypulse.com
servers:
  - url: https://api.citypulse.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /buildings/{id}/sensors:
    get:
      tags:
        - Buildings
        - Sensors
      summary: Get all sensors for a building
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [energy, temperature, humidity, occupancy, security, water_quality, pump, pool, heritage, education, sports, services]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, error]
      responses:
        '200':
          description: List of sensors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sensor'

  # Sensors Management
  /sensors:
    get:
      tags:
        - Sensors
      summary: Get all sensors
      parameters:
        - name: building_id
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of sensors
          content:
            application/json:
              schema:
                type: object
                properties:
                  sensors:
                    type: array
                    items:
                      $ref: '#/components/schemas/Sensor'
                  total:
                    type: integer
                  limit:
                    type: integer
                  offset:
                    type: integer

    post:
      tags:
        - Sensors
      summary: Create new sensor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorCreate'
      responses:
        '201':
          description: Sensor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
  /sensors/{id}:
    get:
      tags:
        - Sensors
      summary: Get sensor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Sensor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'

    patch:
      tags:
        - Sensors
      summary: Update sensor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorUpdate'
      responses:
        '200':
          description: Sensor updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'

    delete:
      tags:
        - Sensors
      summary: Delete sensor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Sensor deleted
  /sensors/{id}/readings:
    get:
      tags:
        - Sensors
        - Energy Data
      summary: Get sensor readings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Sensor readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SensorReading'

    post:
      tags:
        - Sensors
        - Energy Data
      summary: Add sensor reading
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SensorReadingCreate'
      responses:
        '201':
          description: Reading added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SensorReading'

  # Energy Data & Analytics
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Sensor:
      type: object
      properties:
        id:
          type: string
        building_id:
          type: string
        type:
          type: string
          enum: [energy, temperature, humidity, occupancy, security, water_quality, pump, pool, heritage, education, sports, services, renovation, visitors]
        name:
          type: string
        value:
          type: number
        unit:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        last_reading:
          type: string
          format: date-time
        metadata:
          type: object
          properties:
            location:
              type: string
            critical:
              type: boolean
            alert_threshold:
              type: number
            description:
              type: string
    SensorCreate:
      type: object
      required:
        - building_id
        - type
        - name
        - unit
      properties:
        building_id:
          type: string
        type:
          type: string
          enum: [energy, temperature, humidity, occupancy, security, water_quality, pump, pool, heritage, education, sports, services, renovation, visitors]
        name:
          type: string
        unit:
          type: string
        metadata:
          type: object
    SensorUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, inactive, error]
        metadata:
          type: object
    SensorReading:
      type: object
      properties:
        id:
          type: string
        sensor_id:
          type: string
        value:
          type: number
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
    SensorReadingCreate:
      type: object
      required:
        - value
      properties:
        value:
          type: number
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object

    # Energy Data Schemas
...